// src/shared/types/mock-server-data.type.ts
export type MockServerData = {
  title: string[];
  description: string[];
  date: string[];
  city: string[];
  previewImage: string[];
  photo: string[];
  isPremium: string[];
  isFavorite: string[];
  rating: string[];
  type: string[];
  rooms: string[];
  guests: string[];
  price: string[];
  goods: string[];
  user: string[];
  comments: string[];
  location: string[];
};